requirements.txt
# Core Dependencies
agent-framework==1.0.0b260212
python-dotenv

# HTTP Server & Agent Inspector Integration
azure-ai-agentserver-core==1.0.0b10
azure-ai-agentserver-agentframework==1.0.0b10
agent-dev-cli

# Debugging
debugpy

# LLM & API Clients
openai
azure-identity

# Observability & Tracing
opentelemetry-exporter-otlp-proto-grpc

# Image Generation & TTS
dashscope
edge-tts
pydub

# Web UI
streamlit==1.54.0
markdown

# PDF Generation (pure Python, no native dependencies)
fpdf2

# Utilities
requests
pydantic
aiohttp

# Optional: Video & Search Support
qrcode
pillow
ddgs
curl_cffi

# Optional: Development Tools
pytest
black
pylint
